requirements.txt
javascript==1!1.2.6
python-dotenv==1.0.1
requests==2.32.3
